Create native SQLite loader chain
