+++ src/main/java/org/sqlite/bin/SQLiteNativeLoader.java
@@ -1,0 +1,14 @@
+package org.sqlite.bin;
+
+import org.sqlite.NativeLibraryNotFoundException;
+
+/** Loader which loads the SQLite Native binary file. */
+public interface SQLiteNativeLoader {
+
+    /**
+     * Load the SQLite Native binary file.
+     *
+     * @throws NativeLibraryNotFoundException Unable to load the native binary
+     */
+    void loadSQLiteNative() throws NativeLibraryNotFoundException;
+}
